Slides
Here are my slides. My plan is to talk about the different and
intriguing aspects of Miss James (as a way to humanly explain why we
care), how her first names was discovered, and the team data cleanup in
prep for this, all while discussing the TEI. This will likely be shorter
in its final iteration, depending on time.